PCHR-1488: Rename LeavePeriodEntitlement to LeaveBalance
The staff member entitlement is a fixed amount days that they have available
throughout the whole period. Even though the entitlement is stored as
LeaveBalanceChanges records, it will never change. The balance is a more
global term which references the amount of days of the entitlement still available
to be taken as leave. In other words, the balance is:

entitlement - days taken as leave + added days (toil, for example)

Having the leave_period_entitlement as the main point to not only the entitlement,
but to all balance changes could lead to some confusion about the real purpose of
this table and its records, so it's now name leave_balance. The entity methods were
also updated to reflect this change and avoid any confusion between the terms
"entitlement" and "balance"